+++ test/emailpassword/signupFeature.test.js
@@ -412,7 +412,12 @@ describe(`signupFeature: ${printPath("[test/emailpassword/signupFeature.test.js]
                 appName: "SuperTokens",
                 websiteDomain: "supertokens.io",
             },
-            recipeList: [EmailPassword.init(), Session.init()],
+            recipeList: [
+                EmailPassword.init(),
+                Session.init({
+                    enableAntiCsrf: true,
+                }),
+            ],
         });
         const app = express();