+++ test/faunadb.test.js
@@ -63,6 +63,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
                     accessFaunadbTokenFromFrontend: true,
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -179,6 +180,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
                     accessFaunadbTokenFromFrontend: true,
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -298,6 +300,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                 Session.init({
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -379,6 +382,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                 Session.init({
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -491,6 +495,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
                     accessFaunadbTokenFromFrontend: true,
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -583,6 +588,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
                     accessFaunadbTokenFromFrontend: true,
+                    enableAntiCsrf: true,
                 }),
             ],
         });
@@ -730,6 +736,7 @@ describe(`faunaDB: ${printPath("[test/faunadb.test.js]")}`, function () {
                 Session.init({
                     faunadbSecret: "fnAD2HH-Q6ACBSJxMjwU5YT7hvkaVo6Te8PJWqsT",
                     userCollectionName: "users",
+                    enableAntiCsrf: true,
                 }),
             ],
         });