+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2021-01-28
+### Changed
+- enableAntiCsrf as config parameter in session recipe
+- enableAntiCsrf boolean in session create,verify and refresh APIs if CDI version is 2.6
+- cookieSecure to true by default if the apiDomain has https
+- if the apiDomain and websiteDomain values are different (no common top level domain), then cookieSameSite will be set to none by default, else set it to lax
+- Fixed https://github.com/supertokens/supertokens-node/issues/63
+
 ## [3.3.2] - 2021-01-29
 ### Fixed
 - Always sets httpOnly flag to be true for session cookies regardless of if secure flag is on or off.